[SDS-770] fix type movement
