+++ setup.py
@@ -51,8 +51,10 @@ setup(name='qilib',
       description='Quantum Library for the Quantum Inspire platform',
       long_description=get_long_description(),
       long_description_content_type='text/markdown',
+      url='https://qutech.nl',
       version=get_version_number('qilib'),
       author='QuantumInspire',
+      author_email='support@quantum-inspire.com',
       python_requires='>=3.7',
       package_dir={'': 'src'},
       packages=['qilib', 'qilib.configuration_helper', 'qilib.configuration_helper.adapters',
@@ -64,8 +66,7 @@ setup(name='qilib',
           'Programming Language :: Python :: 3.8',
           'Programming Language :: Python :: 3.9',
           'Programming Language :: Python :: 3.10'],
-      license='Other/Proprietary License',
-
+      license='MIT',
       install_requires=['spirack>=0.1.8', 'numpy>=1.20', 'serialize', 'pymongo', 'zhinst;python_version<"3.10"',
                         'requests', 'qcodes==0.33.0', 'qcodes_contrib_drivers==0.13.1', 'dataclasses-json',
                         'zhinst==22.2.26577'],
