+++ src/qilib/version.py
@@ -1,1 +1,1 @@
-__version__ = '0.3.10'
+__version__ = '0.3.11'
